refactor: AnalysisPanel을 상태별 패널 컴포넌트로 분할 + WorkflowContext 도입
800줄짜리 AnalysisPanel을 5개 패널 컴포넌트로 분리:
- DiskImageSection (디스크 이미지 입력)
- StrategyThread (프롬프트 + 전략 카드 + 전략 수정/직접편집)
- McpPlanThread (반려 이력 + MCP 계획 카드 + 반려 카드)
- DoneSummaryPanel (인라인 결과 요약 + 보고서 게이트)
- ChatInputBar (하단 채팅 인풋)

ForensicApp이 AnalysisPanel에 30+ props를 드릴링하던 구조를 WorkflowContext
하나로 정리. AnalysisPanel은 26줄짜리 얇은 오케스트레이터가 되고, 각 패널은
필요한 값만 useWorkflowContext로 골라 쓴다.
